10. Обновление века (часть 1): Создаёт абстрактный класс Smart
